Documented stuff.
